+++ resources/views/users/index.blade.php
@@ -43,7 +43,7 @@
                                         <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
                                     @endcan
                                     @can('user-delete')
-                                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id], 'onsubmit' => 'return confirm("Anda yakin akan menghapus data ini?")', 'style'=>'display:inline']) !!}
                                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                         {!! Form::close() !!}
                                     @endcan
